refactor: move helper out of context
