+++ admin-ui/src/component/my-units/ReservationPopupContent.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { Link } from "react-router-dom";
 import styled from "styled-components";
 import { ReservationType } from "common/types/gql-types";
-import { Permission } from "app/context/permissionHelper";
+import { Permission } from "app/modules/permissionHelper";
 import { reservationUrl } from "../../common/urls";
 import { formatTime } from "../../common/util";
 import { DenseVerticalFlex } from "../../styles/layout";
